Made suggested changes
