+++ test/functional/test_f_multi_keyring.py
@@ -17,6 +17,7 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives.asymmetric import rsa
 
 from aws_encryption_sdk.identifiers import Algorithm, KeyringTraceFlag, WrappingAlgorithm
+from aws_encryption_sdk.internal.defaults import ALGORITHM
 from aws_encryption_sdk.keyring.multi_keyring import MultiKeyring
 from aws_encryption_sdk.keyring.raw_keyring import RawAESKeyring, RawRSAKeyring
 from aws_encryption_sdk.materials_managers import DecryptionMaterials, EncryptionMaterials
@@ -29,22 +30,17 @@ _PROVIDER_ID = "Random Raw Keys"
 _KEY_ID = b"5325b043-5843-4629-869c-64794af77ada"
 _WRAPPING_KEY_AES = b"\xeby-\x80A6\x15rA8\x83#,\xe4\xab\xac`\xaf\x99Z\xc1\xce\xdb\xb6\x0f\xb7\x805\xb2\x14J3"
 
-_SIGNING_KEY = b"aws-crypto-public-key"
-
 _ENCRYPTION_MATERIALS_WITHOUT_DATA_KEY = EncryptionMaterials(
-    algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
-    encryption_context=_ENCRYPTION_CONTEXT,
-    signing_key=_SIGNING_KEY,
+    algorithm=ALGORITHM, encryption_context=_ENCRYPTION_CONTEXT
 )
 
 _ENCRYPTION_MATERIALS_WITH_DATA_KEY = EncryptionMaterials(
-    algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
+    algorithm=ALGORITHM,
     data_encryption_key=RawDataKey(
         key_provider=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
         data_key=b'*!\xa1"^-(\xf3\x105\x05i@B\xc2\xa2\xb7\xdd\xd5\xd5\xa9\xddm\xfae\xa8\\$\xf9d\x1e(',
     ),
     encryption_context=_ENCRYPTION_CONTEXT,
-    signing_key=_SIGNING_KEY,
     keyring_trace=[
         KeyringTrace(
             wrapping_key=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
@@ -124,9 +120,7 @@ def test_multi_keyring_encryption_decryption(multi_keyring, encryption_materials
 
     # Generate decryption materials
     decryption_materials = DecryptionMaterials(
-        algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
-        verification_key=b"ex_verification_key",
-        encryption_context=_ENCRYPTION_CONTEXT,
+        algorithm=ALGORITHM, verification_key=b"ex_verification_key", encryption_context=_ENCRYPTION_CONTEXT
     )
 
     # Call on_decrypt function for the keyring
