+++ test/unit/test_utils.py
@@ -13,223 +13,20 @@
 # language governing permissions and limitations under the License.
 """Test suite for aws_encryption_sdk.internal.utils"""
 import io
-import os
 
 import pytest
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import rsa
 from mock import MagicMock, patch, sentinel
 
 import aws_encryption_sdk.identifiers
 import aws_encryption_sdk.internal.utils
 from aws_encryption_sdk.exceptions import InvalidDataKeyError, SerializationError, UnknownIdentityError
-from aws_encryption_sdk.identifiers import Algorithm, KeyringTraceFlag, WrappingAlgorithm
 from aws_encryption_sdk.internal.defaults import MAX_FRAME_SIZE, MESSAGE_ID_LENGTH
-from aws_encryption_sdk.keyring.multi_keyring import MultiKeyring
-from aws_encryption_sdk.keyring.raw_keyring import Keyring, RawAESKeyring, RawRSAKeyring
-from aws_encryption_sdk.materials_managers import DecryptionMaterials, EncryptionMaterials
-from aws_encryption_sdk.structures import DataKey, EncryptedDataKey, KeyringTrace, MasterKeyInfo, RawDataKey
+from aws_encryption_sdk.structures import DataKey, EncryptedDataKey, MasterKeyInfo, RawDataKey
 
 from .test_values import VALUES
 from .unit_test_utils import assert_prepped_stream_identity
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
-
-
-try:  # Python 3.5.0 and 3.5.1 have incompatible typing modules
-    from typing import Iterable  # noqa pylint: disable=unused-import
-except ImportError:  # pragma: no cover
-    # We only actually need these imports when running the mypy checks
-    pass
-
-
-_ENCRYPTION_CONTEXT = {"encryption": "context", "values": "here"}
-_PROVIDER_ID = "Random Raw Keys"
-_KEY_ID = b"5325b043-5843-4629-869c-64794af77ada"
-_WRAPPING_KEY = b"\xeby-\x80A6\x15rA8\x83#,\xe4\xab\xac`\xaf\x99Z\xc1\xce\xdb\xb6\x0f\xb7\x805\xb2\x14J3"
-_SIGNING_KEY = b"aws-crypto-public-key"
-_DATA_KEY = (
-    b"\x00\xfa\x8c\xdd\x08Au\xc6\x92_4\xc5\xfb\x90\xaf\x8f\xa1D\xaf\xcc\xd25" b"\xa8\x0b\x0b\x16\x92\x91W\x01\xb7\x84"
-)
-_WRAPPING_KEY_AES = b"\xeby-\x80A6\x15rA8\x83#,\xe4\xab\xac`\xaf\x99Z\xc1\xce\xdb\xb6\x0f\xb7\x805\xb2\x14J3"
-
-
-class IdentityKeyring(Keyring):
-    def on_encrypt(self, encryption_materials):
-        # type: (EncryptionMaterials) -> EncryptionMaterials
-        return encryption_materials
-
-    def on_decrypt(self, decryption_materials, encrypted_data_keys):
-        # type: (DecryptionMaterials, Iterable[EncryptedDataKey]) -> DecryptionMaterials
-        return decryption_materials
-
-
-class OnlyGenerateKeyring(Keyring):
-    def on_encrypt(self, encryption_materials):
-        # type: (EncryptionMaterials) -> EncryptionMaterials
-        if encryption_materials.data_encryption_key is None:
-            key_provider = MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID)
-            data_encryption_key = RawDataKey(
-                key_provider=key_provider, data_key=os.urandom(encryption_materials.algorithm.kdf_input_len)
-            )
-            encryption_materials.add_data_encryption_key(
-                data_encryption_key=data_encryption_key,
-                keyring_trace=KeyringTrace(
-                    wrapping_key=key_provider, flags={KeyringTraceFlag.WRAPPING_KEY_GENERATED_DATA_KEY}
-                ),
-            )
-        return encryption_materials
-
-    def on_decrypt(self, decryption_materials, encrypted_data_keys):
-        # type: (DecryptionMaterials, Iterable[EncryptedDataKey]) -> DecryptionMaterials
-        return decryption_materials
-
-
-def get_identity_keyring():
-    return IdentityKeyring()
-
-
-def get_keyring_which_only_generates():
-    return OnlyGenerateKeyring()
-
-
-def get_encryption_materials_with_data_key():
-    return EncryptionMaterials(
-        algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
-        data_encryption_key=RawDataKey(
-            key_provider=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-            data_key=b'*!\xa1"^-(\xf3\x105\x05i@B\xc2\xa2\xb7\xdd\xd5\xd5\xa9\xddm\xfae\xa8\\$\xf9d\x1e(',
-        ),
-        encryption_context=_ENCRYPTION_CONTEXT,
-        signing_key=_SIGNING_KEY,
-        keyring_trace=[
-            KeyringTrace(
-                wrapping_key=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-                flags={KeyringTraceFlag.WRAPPING_KEY_GENERATED_DATA_KEY},
-            )
-        ],
-    )
-
-
-def get_encryption_materials_without_data_key():
-    return EncryptionMaterials(
-        algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
-        encryption_context=_ENCRYPTION_CONTEXT,
-        signing_key=_SIGNING_KEY,
-    )
-
-
-def get_encryption_materials_with_encrypted_data_key():
-    return EncryptionMaterials(
-        algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
-        data_encryption_key=RawDataKey(
-            key_provider=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-            data_key=b'*!\xa1"^-(\xf3\x105\x05i@B\xc2\xa2\xb7\xdd\xd5\xd5\xa9\xddm\xfae\xa8\\$\xf9d\x1e(',
-        ),
-        encrypted_data_keys=[
-            EncryptedDataKey(
-                key_provider=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-                encrypted_data_key=b"\xde^\x97\x7f\x84\xe9\x9e\x98\xd0\xe2\xf8\xd5\xcb\xe9\x7f.}\x87\x16,\x11n#\xc8p"
-                b"\xdb\xbf\x94\x86*Q\x06\xd2\xf5\xdah\x08\xa4p\x81\xf7\xf4G\x07FzE\xde",
-            )
-        ],
-        encryption_context=_ENCRYPTION_CONTEXT,
-        signing_key=_SIGNING_KEY,
-        keyring_trace=[
-            KeyringTrace(
-                wrapping_key=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-                flags={
-                    KeyringTraceFlag.WRAPPING_KEY_GENERATED_DATA_KEY,
-                    KeyringTraceFlag.WRAPPING_KEY_ENCRYPTED_DATA_KEY,
-                },
-            )
-        ],
-    )
-
-
-def get_decryption_materials_with_data_key():
-    return DecryptionMaterials(
-        algorithm=Algorithm.AES_256_GCM_IV12_TAG16_HKDF_SHA384_ECDSA_P384,
-        data_encryption_key=RawDataKey(
-            key_provider=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-            data_key=b'*!\xa1"^-(\xf3\x105\x05i@B\xc2\xa2\xb7\xdd\xd5\xd5\xa9\xddm\xfae\xa8\\$\xf9d\x1e(',
-        ),
-        encryption_context=_ENCRYPTION_CONTEXT,
-        verification_key=b"ex_verification_key",
-        keyring_trace=[
-            KeyringTrace(
-                wrapping_key=MasterKeyInfo(provider_id=_PROVIDER_ID, key_info=_KEY_ID),
-                flags={KeyringTraceFlag.WRAPPING_KEY_DECRYPTED_DATA_KEY},
-            )
-        ],
-    )
-
-
-def get_decryption_materials_without_data_key():
-    return DecryptionMaterials(encryption_context=_ENCRYPTION_CONTEXT, verification_key=b"ex_verification_key")
-
-
-def get_multi_keyring_with_generator_and_children():
-    return MultiKeyring(
-        generator=RawAESKeyring(
-            key_namespace=_PROVIDER_ID,
-            key_name=_KEY_ID,
-            wrapping_algorithm=WrappingAlgorithm.AES_256_GCM_IV12_TAG16_NO_PADDING,
-            wrapping_key=_WRAPPING_KEY_AES,
-        ),
-        children=[
-            RawRSAKeyring(
-                key_namespace=_PROVIDER_ID,
-                key_name=_KEY_ID,
-                wrapping_algorithm=WrappingAlgorithm.RSA_OAEP_SHA256_MGF1,
-                private_wrapping_key=rsa.generate_private_key(
-                    public_exponent=65537, key_size=2048, backend=default_backend()
-                ),
-            ),
-            RawRSAKeyring(
-                key_namespace=_PROVIDER_ID,
-                key_name=_KEY_ID,
-                wrapping_algorithm=WrappingAlgorithm.RSA_OAEP_SHA256_MGF1,
-                private_wrapping_key=rsa.generate_private_key(
-                    public_exponent=65537, key_size=2048, backend=default_backend()
-                ),
-            ),
-        ],
-    )
-
-
-def get_multi_keyring_with_no_children():
-    return MultiKeyring(
-        generator=RawRSAKeyring(
-            key_namespace=_PROVIDER_ID,
-            key_name=_KEY_ID,
-            wrapping_algorithm=WrappingAlgorithm.RSA_OAEP_SHA256_MGF1,
-            private_wrapping_key=rsa.generate_private_key(
-                public_exponent=65537, key_size=2048, backend=default_backend()
-            ),
-        )
-    )
-
-
-def get_multi_keyring_with_no_generator():
-    return MultiKeyring(
-        children=[
-            RawRSAKeyring(
-                key_namespace=_PROVIDER_ID,
-                key_name=_KEY_ID,
-                wrapping_algorithm=WrappingAlgorithm.RSA_OAEP_SHA256_MGF1,
-                private_wrapping_key=rsa.generate_private_key(
-                    public_exponent=65537, key_size=2048, backend=default_backend()
-                ),
-            ),
-            RawAESKeyring(
-                key_namespace=_PROVIDER_ID,
-                key_name=_KEY_ID,
-                wrapping_algorithm=WrappingAlgorithm.AES_128_GCM_IV12_TAG16_NO_PADDING,
-                wrapping_key=_WRAPPING_KEY_AES,
-            ),
-        ]
-    )
 
 
 def test_prep_stream_data_passthrough():
